+++ Configuration/Icons.php
@@ -1,5 +1,9 @@
 <?php
 
+/*
+ * SPDX-License-Identifier: GPL-2.0-or-later
+ */
+
 declare(strict_types=1);
 
 use TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider;